requirements.txt
djongo==1.3.3
dnspython==2.0.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
Pillow
gunicorn